feat: add access level control to application

+++ app/Http/Controllers/API/UserController.php
@@ -20,6 +20,7 @@ class UserController extends Controller
      */
     public function index()
     {
+        $this->authorize('isAdmin');
         return User::latest()->paginate(10);
     }
 
@@ -105,7 +106,8 @@ class UserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {
+    {   
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
         $user->delete();
         return ['meassage' => 'User Deleted'];
